Update list.md (#1122)
* Update list.md

Add clarifying statement on the significance for using the special is="" attribute. This may be useful for newbies new to web development and/or front-end frameworks. The is="" attribute isn't a well-known html attribute outside of the web components community.

* Update list.md

Add clarifying statement on use of $emit method since first time being used up until this point in the docs.

* Tweak wording for note about "is" attribute